# pref: implement readBuffer to eliminate extra copying