Benchmark:
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/decoder
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
         │   old.txt    │              new.txt               │
         │    sec/op    │   sec/op     vs base               │
Decode-4   105.75m ± 1%   99.00m ± 3%  -6.39% (p=0.000 n=10)

         │   old.txt    │            new.txt             │
         │     B/op     │     B/op      vs base          │
Decode-4   73.49Mi ± 0%   73.49Mi ± 0%  ~ (p=0.050 n=10)

         │   old.txt   │             new.txt             │
         │  allocs/op  │  allocs/op   vs base            │
Decode-4   100.0k ± 0%   100.0k ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```

close #179